Split code into main.rs and lib.rs
Basically all the code is now in lib.rs. In main.rs only the highest
level setup is done.

Also included the following improvements:

- Locate the font file from the cargo project root
- Use a system-wide font file as a fallback
- Do not draw a grid by default, but can be toggled by pressing `G`
- Added Vim-style interface for the menu navigation
